+++ forge/src/main/java/de/cech12/unlitcampfire/platform/ForgePlatformHelper.java
@@ -2,6 +2,7 @@ package de.cech12.unlitcampfire.platform;
 
 import de.cech12.unlitcampfire.platform.services.IPlatformHelper;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ItemStackTemplate;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.loading.FMLLoader;
@@ -18,7 +19,7 @@ public class ForgePlatformHelper implements IPlatformHelper {
 
     @Override
     public boolean isModLoaded(String modId) {
-        return ModList.get().isLoaded(modId);
+        return ModList.isLoaded(modId);
     }
 
     @Override
@@ -32,8 +33,8 @@ public class ForgePlatformHelper implements IPlatformHelper {
     }
 
     @Override
-    public ItemStack getRemainingStackAfterUsage(ItemStack usedStack) {
-        return usedStack.getItem().getCraftingRemainder(usedStack);
+    public ItemStackTemplate getRemainingStackAfterUsage(ItemStack usedStack) {
+        return usedStack.getCraftingRemainder();
     }
 
 }