+++ neoforge/src/main/java/de/cech12/unlitcampfire/platform/NeoForgeConfigHelper.java
@@ -21,24 +21,24 @@ public class NeoForgeConfigHelper implements IConfigHelper {
 
     private static final ModConfigSpec SERVER_CONFIG;
 
-    public static final ModConfigSpec.IntValue CAMPFIRE_LIT_TIME;
-    public static final ModConfigSpec.IntValue CAMPFIRE_RUN_OUT_INDICATOR_TIME;
-    public static final ModConfigSpec.IntValue CAMPFIRE_RAIN_UNLIT_TIME;
+    public static final ModConfigSpec.LongValue CAMPFIRE_LIT_TIME;
+    public static final ModConfigSpec.LongValue CAMPFIRE_RUN_OUT_INDICATOR_TIME;
+    public static final ModConfigSpec.LongValue CAMPFIRE_RAIN_UNLIT_TIME;
     public static final ModConfigSpec.IntValue CAMPFIRE_RAIN_PARTICLE_FACTOR;
     public static final ModConfigSpec.BooleanValue CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME;
     public static final ModConfigSpec.BooleanValue CAMPFIRE_ADDING_BURNABLES;
-    public static final ModConfigSpec.IntValue CAMPFIRE_MAX_LIT_TIME_EXTENSION;
+    public static final ModConfigSpec.LongValue CAMPFIRE_MAX_LIT_TIME_EXTENSION;
     public static final ModConfigSpec.BooleanValue CAMPFIRE_AFFECTED_BY_SLEEP_TIME;
     public static final ModConfigSpec.BooleanValue GENERATED_CAMPFIRE_IS_LIT_INFINITELY;
     public static final ModConfigSpec.BooleanValue INFINITE_CAMPFIRE_IGNORES_RAIN;
 
-    public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_LIT_TIME;
-    public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME;
-    public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_RAIN_UNLIT_TIME;
+    public static final ModConfigSpec.LongValue SOUL_CAMPFIRE_LIT_TIME;
+    public static final ModConfigSpec.LongValue SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME;
+    public static final ModConfigSpec.LongValue SOUL_CAMPFIRE_RAIN_UNLIT_TIME;
     public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR;
     public static final ModConfigSpec.BooleanValue SOUL_CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME;
     public static final ModConfigSpec.BooleanValue SOUL_CAMPFIRE_ADDING_BURNABLES;
-    public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION;
+    public static final ModConfigSpec.LongValue SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION;
     public static final ModConfigSpec.BooleanValue SOUL_CAMPFIRE_AFFECTED_BY_SLEEP_TIME;
     public static final ModConfigSpec.BooleanValue GENERATED_SOUL_CAMPFIRE_IS_LIT_INFINITELY;
     public static final ModConfigSpec.BooleanValue INFINITE_SOUL_CAMPFIRE_IGNORES_RAIN;
@@ -139,7 +139,7 @@ public class NeoForgeConfigHelper implements IConfigHelper {
     }
 
     @Override
-    public int getLitTime(boolean isSoulCampfire) {
+    public long getLitTime(boolean isSoulCampfire) {
         try {
             return isSoulCampfire ? SOUL_CAMPFIRE_LIT_TIME.get() : CAMPFIRE_LIT_TIME.get();
         } catch (IllegalStateException ex) {
@@ -148,7 +148,7 @@ public class NeoForgeConfigHelper implements IConfigHelper {
     }
 
     @Override
-    public int getRunOutIndicatorTime(boolean isSoulCampfire) {
+    public long getRunOutIndicatorTime(boolean isSoulCampfire) {
         try {
             return isSoulCampfire ? SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME.get() : CAMPFIRE_RUN_OUT_INDICATOR_TIME.get();
         } catch (IllegalStateException ex) {
@@ -157,7 +157,7 @@ public class NeoForgeConfigHelper implements IConfigHelper {
     }
 
     @Override
-    public int getRainUnlitTime(boolean isSoulCampfire) {
+    public long getRainUnlitTime(boolean isSoulCampfire) {
         try {
             return isSoulCampfire ? SOUL_CAMPFIRE_RAIN_UNLIT_TIME.get() : CAMPFIRE_RAIN_UNLIT_TIME.get();
         } catch (IllegalStateException ex) {
@@ -193,7 +193,7 @@ public class NeoForgeConfigHelper implements IConfigHelper {
     }
 
     @Override
-    public int getMaxLitTimeExtension(boolean isSoulCampfire) {
+    public long getMaxLitTimeExtension(boolean isSoulCampfire) {
         try {
             return isSoulCampfire ? SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION.get() : CAMPFIRE_MAX_LIT_TIME_EXTENSION.get();
         } catch (IllegalStateException ex) {